+++ src/Core/Dfe.Complete.Application/Projects/Commands/UpdateProject/UpdateSignificantDate.cs
@@ -3,6 +3,7 @@ using Dfe.Complete.Domain.Enums;
 using Dfe.Complete.Domain.Interfaces.Repositories;
 using Dfe.Complete.Domain.ValueObjects;
 using Dfe.Complete.Utils;
+using Dfe.Complete.Utils.Exceptions;
 using MediatR;
 
 namespace Dfe.Complete.Application.Projects.Commands.UpdateProject;