+++ src/Core/Dfe.Complete.Application/Projects/Queries/GetProject/GetProjectById.cs
@@ -6,7 +6,7 @@ using Dfe.Complete.Application.Projects.Models;
 using Dfe.Complete.Application.Projects.Interfaces;
 using Dfe.Complete.Application.Projects.Queries.QueryFilters;
 using Microsoft.EntityFrameworkCore;
-using Dfe.Complete.Utils;
+using Dfe.Complete.Utils.Exceptions;
 
 namespace Dfe.Complete.Application.Projects.Queries.GetProject
 {